+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.97.2", date: "2026-03-22" };
+  return { version: "8.97.3", date: "2026-03-22" };
 }